backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
python-dotenv==1.0.0
sendgrid==6.11.0
imagekitio==4.1.0
pydantic[email]>=2.0.0
websockets==12.0
scikit-learn>=1.3.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
python-jose[cryptography]==3.3.0
psutil==5.9.6
pytesseract==0.3.10
Pillow==10.1.0
aiohttp==3.9.1
requests==2.31.0
reportlab==4.0.7
google-generativeai==0.3.2